show comments thing

+++ app/Http/Controllers/UsersController.php
@@ -109,7 +109,7 @@ class UsersController extends Controller
                     }
                     
                     $filesCount = File::whereIn('user_id', $ids)->count();
-                    $creditsCount = Credit::whereIn('user_id', $ids)
+                    $creditsCount = (int) Credit::whereIn('user_id', $ids)
                     ->where('credits', '>', 0)->sum('credits');
 
                 }
@@ -130,7 +130,7 @@ class UsersController extends Controller
                     }
                     
                     $filesCount = File::whereIn('user_id', $ids)->count();
-                    $creditsCount = Credit::whereIn('user_id', $ids)
+                    $creditsCount = (int) Credit::whereIn('user_id', $ids)
                     ->where('credits', '>', 0)->sum('credits');
 
                 }
